# openapi.yaml
...
openapi: 3.0.3
info:
  title: Oficina MVP API
  description: Tech Challenge 15SOAT — Sistema de Gestão de Oficina Mecânica (fases
    1 e 2).
  version: 1.0.0
...
paths:
...
  /ordens-servico/{id}/pecas/{itemId}:
    delete:
      tags:
      - Ordens de Serviço
      summary: Remove item de peça (somente antes de gerar orçamento)
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      - name: itemId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdemServicoResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
...
  /ordens-servico/{id}/servicos/{itemId}:
    delete:
      tags:
      - Ordens de Serviço
      summary: Remove item de serviço (somente antes de gerar orçamento)
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      - name: itemId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdemServicoResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
...
  /pecas:
    get:
      tags:
      - Peças / Estoque
      summary: Lista todas as peças
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PecaResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
    post:
      tags:
      - Peças / Estoque
      summary: Cadastra uma nova peça
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CadastrarPecaRequest"
      responses:
        "201":
          description: Peça criada
        "400":
          description: Payload inválido
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
  /pecas/{id}:
    get:
      tags:
      - Peças / Estoque
      summary: Busca uma peça pelo ID (inclui saldo disponível)
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "404":
          description: Peça não encontrada
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
    put:
      tags:
      - Peças / Estoque
      summary: Atualiza dados de catálogo da peça (descrição e valor unitário)
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AtualizarPecaRequest"
      responses:
        "404":
          description: Peça não encontrada
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
    delete:
      tags:
      - Peças / Estoque
      summary: Remove uma peça (bloqueado se houver reservas ativas)
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "204":
          description: Removida
        "404":
          description: Peça não encontrada
        "409":
          description: Peça possui reservas ativas
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
  /pecas/{id}/saldo:
    post:
      tags:
      - Peças / Estoque
      summary: Adiciona saldo (compra/devolução) à peça
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdicionarSaldoRequest"
      responses:
        "200":
          description: Saldo atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PecaResponse"
        "400":
          description: Quantidade inválida
        "404":
          description: Peça não encontrada
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
  /publico/ordens-servico/{id}:
    get:
      tags:
      - Público
      summary: Consulta resumida de uma OS para acompanhamento do cliente
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "404":
          description: OS não encontrada
  /publico/ordens-servico/{id}/orcamento/decisao:
    post:
      tags:
      - Público
      summary: Recebe notificação externa de aprovação ou recusa do orçamento pelo
        cliente
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecisaoOrcamentoRequest"
      responses:
        "200":
          description: Decisão registrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusOSResponse"
        "404":
          description: OS não encontrada
        "409":
          description: OS não está aguardando aprovação
  /publico/ordens-servico/{id}/status:
    get:
      tags:
      - Público
      summary: "Consulta de status da OS: situação atual com descrição amigável"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "404":
          description: OS não encontrada
...
  /servicos:
    get:
      tags:
      - Serviços
      summary: Lista todos os serviços
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServicoResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
    post:
      tags:
      - Serviços
      summary: Cadastra um novo serviço
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CadastrarServicoRequest"
      responses:
        "201":
          description: Serviço criado
        "400":
          description: Payload inválido
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
  /servicos/{id}:
    get:
      tags:
      - Serviços
      summary: Busca um serviço pelo ID
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "404":
          description: Serviço não encontrado
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
    put:
      tags:
      - Serviços
      summary: Atualiza um serviço
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AtualizarServicoRequest"
      responses:
        "404":
          description: Serviço não encontrado
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
    delete:
      tags:
      - Serviços
      summary: Remove um serviço
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "204":
          description: Removido
        "404":
          description: Serviço não encontrado
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
...
components:
  schemas:
...
    AdicionarSaldoRequest:
      required:
      - quantidade
      type: object
      properties:
        quantidade:
          format: int32
          description: Quantidade de peças a adicionar ao saldo
          minimum: 0
          exclusiveMinimum: true
          type: integer
          example: 10
...
    AtualizarPecaRequest:
      required:
      - descricao
      - valorUnitario
      type: object
      properties:
        descricao:
          description: Descrição da peça
          maxLength: 200
          pattern: \S
          type: string
          example: Filtro de Óleo
        valorUnitario:
          description: Valor unitário da peça
          minimum: 0
          pattern: "^\\d{1,10}([.]\\d{1,2})?$"
          type: number
          example: 45.9
    AtualizarServicoRequest:
      required:
      - descricao
      - valorBase
      type: object
      properties:
        descricao:
          description: Descrição do serviço
          maxLength: 200
          pattern: \S
          type: string
          example: Troca de Óleo
        valorBase:
          description: Valor base do serviço
          minimum: 0
          pattern: "^\\d{1,10}([.]\\d{1,2})?$"
          type: number
          example: 120.0
...
    CadastrarPecaRequest:
      required:
      - descricao
      - valorUnitario
      type: object
      properties:
        descricao:
          description: Descrição da peça
          maxLength: 200
          pattern: \S
          type: string
          example: Filtro de Óleo
        valorUnitario:
          description: Valor unitário da peça
          minimum: 0
          pattern: "^\\d{1,10}([.]\\d{1,2})?$"
          type: number
          example: 45.9
        quantidadeInicial:
          format: int32
          description: Quantidade inicial em estoque
          minimum: 0
          type: integer
          example: 10
    CadastrarServicoRequest:
      required:
      - descricao
      - valorBase
      type: object
      properties:
        descricao:
          description: Descrição do serviço
          maxLength: 200
          pattern: \S
          type: string
          example: Troca de Óleo
        valorBase:
          description: Valor base do serviço
          minimum: 0
          pattern: "^\\d{1,10}([.]\\d{1,2})?$"
          type: number
          example: 120.0
...
    DecisaoOrcamentoRequest:
      required:
      - aprovado
      type: object
      properties:
        aprovado:
          description: true = orçamento aprovado; false = recusado
          type: boolean
          example: true
...
    ItemPecaView:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        pecaId:
          $ref: "#/components/schemas/UUID"
        quantidade:
          format: int32
          type: integer
        valorUnitario:
          type: number
        reservaId:
          $ref: "#/components/schemas/UUID"
...
    ItemServicoView:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        servicoId:
          $ref: "#/components/schemas/UUID"
        valorCobrado:
          type: number
...
    OffsetDateTime:
      format: date-time
      type: string
      example: 2022-03-10T12:15:50-04:00
    OrcamentoView:
      type: object
      properties:
        valorTotal:
          type: number
        geradoEm:
          $ref: "#/components/schemas/OffsetDateTime"
        aprovadoEm:
          $ref: "#/components/schemas/OffsetDateTime"
    OrdemServicoResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        clienteId:
          $ref: "#/components/schemas/UUID"
        veiculoId:
          $ref: "#/components/schemas/UUID"
        status:
          type: string
        itensServico:
          type: array
          items:
            $ref: "#/components/schemas/ItemServicoView"
        itensPeca:
          type: array
          items:
            $ref: "#/components/schemas/ItemPecaView"
        orcamento:
          $ref: "#/components/schemas/OrcamentoView"
        criadaEm:
          $ref: "#/components/schemas/OffsetDateTime"
        diagnosticoIniciadoEm:
          $ref: "#/components/schemas/OffsetDateTime"
        execucaoIniciadaEm:
          $ref: "#/components/schemas/OffsetDateTime"
        finalizadaEm:
          $ref: "#/components/schemas/OffsetDateTime"
        entregueEm:
          $ref: "#/components/schemas/OffsetDateTime"
        canceladaEm:
          $ref: "#/components/schemas/OffsetDateTime"
    PecaResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        descricao:
          type: string
        valorUnitario:
          type: number
        quantidadeTotal:
          format: int32
          type: integer
        saldoDisponivel:
          format: int32
          type: integer
...
    ServicoResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        descricao:
          type: string
        valorBase:
          type: number
    StatusOSResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        status:
          type: string
        descricao:
          type: string
        atualizadoEm:
          $ref: "#/components/schemas/OffsetDateTime"
...
    UUID:
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      description: "JWT emitido pelo POST /auth/login (cole apenas o token, sem prefixo)."
      scheme: bearer
      bearerFormat: JWT
    SecurityScheme:
      type: http
      description: Authentication
      scheme: bearer
      bearerFormat: JWT